+++ Sources/SwiftUIKit/_Deprecated/ForEachMultiPicker.swift
@@ -8,14 +8,7 @@
 
 import SwiftUI
 
-/**
- This generic picker lists `Identifiable` items in a SwiftUI
- `ForEach` and binds its `selection` to an external value.
- 
- You can use this view instead of the native SwiftUI `Picker`
- to get more control over the list item views. The view uses
- the provided `listItem` to build an item view for each item.
- */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PickerKit")
 public struct ForEachMultiPicker<Item: Identifiable, ItemView: View>: View {
     
     /// Create a for-each multi-picker.
@@ -50,9 +43,6 @@ public struct ForEachMultiPicker<Item: Identifiable, ItemView: View>: View {
             })
         }
     }
-}
-
-private extension ForEachMultiPicker {
     
     var selectedIds: [Item.ID] {
         selection.wrappedValue.map { $0.id }
@@ -81,55 +71,4 @@ private extension View {
         self
         #endif
     }
-}
-
-#Preview {
-    
-    struct Preview: View {
-        
-        @State
-        private var selection = [PreviewItem.all[0]]
-        
-        var body: some View {
-            NavigationView {
-                List {
-                    ForEachMultiPicker(
-                        items: PreviewItem.all,
-                        selection: $selection) { item, isSelected in
-                            ListSelectItem(isSelected: isSelected) {
-                                Text(item.name)
-                            }
-                        }
-                }
-                .withTitle("Pick multiple items")
-            }
-        }
-    }
-    
-    struct PreviewItem: Identifiable, Equatable {
-        
-        let name: String
-        
-        var id: String { name }
-        
-        static let all = [
-            PreviewItem(name: "Item #1"),
-            PreviewItem(name: "Item #2"),
-            PreviewItem(name: "Item #3"),
-            PreviewItem(name: "Item #4"),
-            PreviewItem(name: "Item #5"),
-            PreviewItem(name: "Item #6"),
-            PreviewItem(name: "Item #7"),
-            PreviewItem(name: "Item #8"),
-            PreviewItem(name: "Item #9"),
-            PreviewItem(name: "Item #10"),
-            PreviewItem(name: "Item #11"),
-            PreviewItem(name: "Item #12"),
-            PreviewItem(name: "Item #13"),
-            PreviewItem(name: "Item #14"),
-            PreviewItem(name: "Item #15")
-        ]
-    }
-    
-    return Preview()
 }