+++ Sources/SwiftUIKit/_Deprecated/ForEachPicker.swift
@@ -8,17 +8,7 @@
 
 import SwiftUI
 
-/**
- This generic picker lists `Identifiable` items in a SwiftUI
- `ForEach` and binds its `selection` to an external value.
- 
- You can use this view instead of the native SwiftUI `Picker`
- to get more control over the list item views. The view uses
- the provided `listItem` to build an item view for each item.
- 
- If `dismissAfterPick` is `true` the picker dismisses itself
- automatically when an item is picked.
- */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/PickerKit")
 public struct ForEachPicker<Item: Identifiable, ItemView: View>: View {
     
     /// Create a for-each picker.
@@ -61,9 +51,6 @@ public struct ForEachPicker<Item: Identifiable, ItemView: View>: View {
             })
         }
     }
-}
-
-private extension ForEachPicker {
     
     var selectedId: Item.ID {
         selection.wrappedValue.id
@@ -105,54 +92,4 @@ private extension View {
         self
         #endif
     }
-}
-
-#Preview {
-    
-    struct Preview: View {
-        
-        @State private var selection = PreviewItem.all[0]
-        
-        var body: some View {
-            NavigationView {
-                List {
-                    ForEachPicker(
-                        items: PreviewItem.all,
-                        selection: $selection) { item, isSelected in
-                            ListSelectItem(isSelected: isSelected) {
-                                Text(item.name)
-                            }
-                        }
-                }
-                .withTitle("Pick an item")
-            }
-        }
-    }
-    
-    struct PreviewItem: Identifiable, Equatable {
-        
-        let name: String
-        
-        var id: String { name }
-        
-        static let all = [
-            PreviewItem(name: "Item #1"),
-            PreviewItem(name: "Item #2"),
-            PreviewItem(name: "Item #3"),
-            PreviewItem(name: "Item #4"),
-            PreviewItem(name: "Item #5"),
-            PreviewItem(name: "Item #6"),
-            PreviewItem(name: "Item #7"),
-            PreviewItem(name: "Item #8"),
-            PreviewItem(name: "Item #9"),
-            PreviewItem(name: "Item #10"),
-            PreviewItem(name: "Item #11"),
-            PreviewItem(name: "Item #12"),
-            PreviewItem(name: "Item #13"),
-            PreviewItem(name: "Item #14"),
-            PreviewItem(name: "Item #15")
-        ]
-    }
-    
-    return Preview()
 }