+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ SwiftUIKit makes its best effort to honor semver, but breaking changes can occur
 * `DocumentCamera` has been moved to https://github.com/danielsaidi/PickerKit.
 * `ImagePicker` has been moved to https://github.com/danielsaidi/PickerKit.
 * `FilePicker` has been moved to https://github.com/danielsaidi/PickerKit.
+* `ForEachMultiPicker` has been moved to https://github.com/danielsaidi/PickerKit.
 * `OptionalBinding` has been moved to https://github.com/danielsaidi/PickerKit.
 * `PhotoCamera` has been moved to https://github.com/danielsaidi/PickerKit.